Fix stub file, permission

+++ resources/views/datatable-actions/role.blade.php
@@ -1,8 +1,13 @@
 <div class="center action-buttons">
-    @if ($query->name !== 'superadmin')
-        <a href="{{ route('role.permission', $query->id) }}" title="Set Permission" class="btn btn-sm btn-outline-primary" data-id="{{ $query->id }}"><i class="ti ti-lock"></i> Set Access Menu & Permission</a>
+    @if ((auth()->user()->hasRole('superadmin')) | (auth()->user()->hasPermissionTo('edit role')))
+        @if ($query->name !== 'superadmin')
+            <a href="{{ route('role.permission', $query->id) }}" title="Set Permission" class="btn btn-sm btn-outline-primary" data-id="{{ $query->id }}"><i class="ti ti-lock"></i> Set Access Menu & Permission</a>
+        @endif
+        
+        <x-buttons.button-primary title="Edit" class="btn-sm btn-edit" data-id="{{ $query->id }}"><i class="ti ti-pencil"></i></x-buttons.button-primary>
     @endif
-    <x-buttons.button-primary title="Edit" class="btn-sm btn-edit" data-id="{{ $query->id }}"><i class="ti ti-pencil"></i></x-buttons.button-primary>
-    <x-buttons.button-danger title="Delete" class="btn-sm btn-delete"><i class="ti ti-trash"></i></x-buttons.button-danger>
-    <x-inputs.form action="{{ route('role.delete', $query->id) }}" method="POST" class="form-delete">@method('DELETE')</x-inputs.form>
+    @if ((auth()->user()->hasRole('superadmin')) | (auth()->user()->hasPermissionTo('delete role')))
+        <x-buttons.button-danger title="Delete" class="btn-sm btn-delete"><i class="ti ti-trash"></i></x-buttons.button-danger>
+        <x-inputs.form action="{{ route('role.delete', $query->id) }}" method="POST" class="form-delete">@method('DELETE')</x-inputs.form>
+    @endif
 </div>
